Refactor "globals" package to "api"
