+++ src/main/java/net/lukkit/lukkit/sandbox/api/HTTP.java
@@ -1,0 +1,5 @@
+package net.lukkit.lukkit.sandbox.api;
+
+public class HTTP {
+    // TODO
+}
