+++ src/main/java/net/lukkit/lukkit/sandbox/api/Plugin.java
@@ -1,0 +1,4 @@
+package net.lukkit.lukkit.sandbox.api;
+
+public class Plugin {
+}
