--- src/main/java/net/lukkit/lukkit/sandbox/globals/HTTP.java
@@ -1,5 +1,0 @@
-package net.lukkit.lukkit.sandbox.globals;
-
-public class HTTP {
-    // TODO
-}
