--- src/main/java/net/lukkit/lukkit/sandbox/globals/Plugin.java
@@ -1,4 +1,0 @@
-package net.lukkit.lukkit.sandbox.globals;
-
-public class Plugin {
-}
